In all the methods used for reading or writing memory, use uint32_t for the size arguments instead of size_t. 1. We won't be reading and writing gigabytes or more. 2. I don't want the size of the type to change based on the instruction set.

+++ branches/mago64/DebugEngine/Exec/DebuggerProxy.h
@@ -48,16 +48,16 @@ namespace MagoCore
         HRESULT ReadMemory( 
             IProcess* process, 
             Address address,
-            SIZE_T length, 
-            SIZE_T& lengthRead, 
-            SIZE_T& lengthUnreadable, 
+            uint32_t length, 
+            uint32_t& lengthRead, 
+            uint32_t& lengthUnreadable, 
             uint8_t* buffer );
 
         HRESULT WriteMemory( 
             IProcess* process, 
             Address address,
-            SIZE_T length, 
-            SIZE_T& lengthWritten, 
+            uint32_t length, 
+            uint32_t& lengthWritten, 
             uint8_t* buffer );
 
         HRESULT SetBreakpoint( IProcess* process, Address address );
