+++ branches/mago64/DebugEngine/Exec/Exec.h
@@ -88,7 +88,7 @@ class Exec
     DEBUG_EVENT     mLastEvent;
 
     wchar_t*        mPathBuf;
-    size_t          mPathBufLen;
+    uint32_t        mPathBufLen;
 
     ProcessMap*     mProcMap;
     PathResolver*   mResolver;
@@ -166,9 +166,9 @@ public:
     HRESULT ReadMemory( 
         IProcess* process, 
         Address address,
-        SIZE_T length, 
-        SIZE_T& lengthRead, 
-        SIZE_T& lengthUnreadable, 
+        uint32_t length, 
+        uint32_t& lengthRead, 
+        uint32_t& lengthUnreadable, 
         uint8_t* buffer );
 
     // Writes a block of memory to a process's address space. The memory is 
@@ -177,8 +177,8 @@ public:
     HRESULT WriteMemory(
         IProcess* process,
         Address address,
-        SIZE_T length,
-        SIZE_T& lengthWritten, 
+        uint32_t length,
+        uint32_t& lengthWritten, 
         uint8_t* buffer );
 
     // Adds or removes a breakpoint. If the process is running when this
