+++ branches/mago64/DebugEngine/MagoNatDE/MemoryBytes.cpp
@@ -41,10 +41,10 @@ namespace Mago
         if ( (rgbMemory == NULL) || (pdwRead == NULL) )
             return E_INVALIDARG;
 
-        HRESULT hr = S_OK;
-        Address addr = 0;
-        SIZE_T  lenRead = 0;
-        SIZE_T  lenUnreadable = 0;
+        HRESULT     hr = S_OK;
+        Address     addr = 0;
+        uint32_t    lenRead = 0;
+        uint32_t    lenUnreadable = 0;
         CComQIPtr<IMagoMemoryContext>   memCxt = pStartContext;
 
         if ( memCxt == NULL )
@@ -80,9 +80,9 @@ namespace Mago
         if ( rgbMemory == NULL )
             return E_INVALIDARG;
 
-        HRESULT hr = S_OK;
-        Address addr = 0;
-        SIZE_T  lenWritten = 0;
+        HRESULT     hr = S_OK;
+        Address     addr = 0;
+        uint32_t    lenWritten = 0;
         CComQIPtr<IMagoMemoryContext>   memCxt = pStartContext;
 
         if ( memCxt == NULL )
